synchro/requirements.txt
beautifulsoup4==4.6.0
certifi==2017.4.17
chardet==3.0.4
Django==1.11.17
django-appconf==1.0.2
django-compressor==2.1.1
django-modelcluster==3.1
django-overextends==0.4.3
django-taggit==0.22.1
django-treebeard==4.1.2
djangorestframework==3.6.3
html5lib==0.999999999
idna==2.5
olefile==0.44
Pillow==4.2.1
pytz==2017.2
rcssmin==1.0.6
requests==2.22.0
rjsmin==1.0.12
six==1.10.0
stripe==2.37.2
Unidecode==0.4.21
urllib3==1.21.1
wagtail>=1.12,<1.13
webencodings==0.5.1
Willow==0.4
